refactor(platform): 合并 notification 和 system_utils 为 platform.py
- 创建 runtime/infra/platform.py，整合系统级服务
- 移动 send_windows_notification 和 KeepSystemAwake
- 删除 ui/common/ 和 controller/system_utils.py
- 更新所有导入路径
