+++ src/danmaku_sender/controller/monitor_controller.py
@@ -4,8 +4,7 @@ import threading
 from PySide6.QtCore import QObject, Signal, Slot
 
 from .concurrency import WorkerThread
-from .system_utils import KeepSystemAwake
-
+from danmaku_sender.runtime.infra.platform import KeepSystemAwake
 from danmaku_sender.repo.bili_api_client import BiliApiClient
 from danmaku_sender.repo.history_manager import HistoryManager
 from danmaku_sender.types.models.common import VideoTarget
